refactor(graph): 拆分人物关系图页面结构与画布逻辑
